requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116  # for GPU versions of torch, torchvision
aim==3.17.5
gputil==1.4.0
gymnasium==0.26.3
hydra-core==1.3.2
memray==1.7.0
protobuf==3.20.3
py-spy==0.3.14
simfire==1.5.3
tensorboardX==2.6
tensorflow-probability==0.19.0
# specifying explicit plus-notation below so pip overwrites cpu verisons (if they exist)
torch==1.13.0+cu116
torchvision==0.14.0+cu116
